Persistance Hierarchy
Inserted the rough template for the hierarchy of data retrieval and storage.

This is actually a pretty big job that needs both our attention now that I work through it.

But it shouldn't be too bad because we have all the neccesary code there. Just need a bit of logic input from us and rearranging code.